Continue creating Queries, started creating a own class for player in order to understand more the code, changed the methods to entities and changed the logic for those methods so now we can use a virtual data base called h2, set names for some values in players, and continue working on it

+++ src/main/java/com/iwaconsolti/league/demo/dto/PlayerDTO.java
@@ -8,12 +8,12 @@ import lombok.ToString;
 @Setter
 @Getter
 public class PlayerDTO {
-    private int id;
+    private long id;
     private String name;
     private String team;
 
 
-    public PlayerDTO(int id, String name, String team) {
+    public PlayerDTO(long id, String name, String team) {
         this.id = id;
         this.name = name;
         this.team = team;
